tests/wpt/tests/css/css-fonts/font-size-sign-function.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Fonts test: font-size with CSS sign() function</title>
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-size-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
:root {
  font-size: 16px;
}
</style>
<div id="target"></div>
<script>
  test_computed_value("font-size", "calc(sign(1rem - 1px) * 100%)", "16px");
</script>
